refactor: extract common code parquet table for adding arrow-rs impl.
