+++ src/query/storages/parquet/src/parquet2/parquet_reader/deserialize.rs
@@ -32,116 +32,16 @@ use common_exception::ErrorCode;
 use common_exception::Result;
 use common_expression::DataBlock;
 use common_expression::DataSchema;
-use common_expression::FieldIndex;
 use common_storage::ColumnNode;
 
 use super::filter::FilterState;
+use crate::parquet2::parquet_reader::Parquet2Reader;
 use crate::parquet_part::ColumnMeta;
-use crate::parquet_part::ParquetRowGroupPart;
-use crate::parquet_reader::ParquetReader;
 
-impl ParquetReader {
-    pub fn deserialize(
-        &self,
-        part: &ParquetRowGroupPart,
-        chunks: Vec<(FieldIndex, Vec<u8>)>,
-        filter: Option<Bitmap>,
-    ) -> Result<DataBlock> {
-        if chunks.is_empty() {
-            return Ok(DataBlock::new(vec![], part.num_rows));
-        }
-
-        let mut chunk_map: HashMap<FieldIndex, Vec<u8>> = chunks.into_iter().collect();
-        let mut columns_array_iter = Vec::with_capacity(self.projected_arrow_schema.fields.len());
-        let mut nested_columns_array_iter =
-            Vec::with_capacity(self.projected_arrow_schema.fields.len());
-        let mut normal_fields = Vec::with_capacity(self.projected_arrow_schema.fields.len());
-        let mut nested_fields = Vec::with_capacity(self.projected_arrow_schema.fields.len());
-
-        let column_nodes = &self.projected_column_nodes.column_nodes;
-        let mut cnt_map = Self::build_projection_count_map(column_nodes);
-
-        for (idx, column_node) in column_nodes.iter().enumerate() {
-            let indices = &column_node.leaf_indices;
-            let mut metas = Vec::with_capacity(indices.len());
-            let mut chunks = Vec::with_capacity(indices.len());
-            for index in indices {
-                // in `read_parquet` function, there is no `TableSchema`, so index treated as column id
-                let column_meta = &part.column_metas[index];
-                let cnt = cnt_map.get_mut(index).unwrap();
-                *cnt -= 1;
-                let column_chunk = if cnt > &mut 0 {
-                    chunk_map.get(index).unwrap().clone()
-                } else {
-                    chunk_map.remove(index).unwrap()
-                };
-                let descriptor = &self.projected_column_descriptors[index];
-                metas.push((column_meta, descriptor));
-                chunks.push(column_chunk);
-            }
-            if let Some(ref bitmap) = filter {
-                // Filter push down for nested type is not supported now.
-                // If the array is nested type, do not push down filter to it.
-                if chunks.len() > 1 {
-                    nested_columns_array_iter.push(Self::to_array_iter(
-                        metas,
-                        chunks,
-                        part.num_rows,
-                        column_node.field.clone(),
-                    )?);
-                    nested_fields.push(self.output_schema.field(idx).clone());
-                } else {
-                    columns_array_iter.push(Self::to_array_iter_with_filter(
-                        metas,
-                        chunks,
-                        part.num_rows,
-                        column_node.field.clone(),
-                        bitmap.clone(),
-                    )?);
-                    normal_fields.push(self.output_schema.field(idx).clone());
-                }
-            } else {
-                columns_array_iter.push(Self::to_array_iter(
-                    metas,
-                    chunks,
-                    part.num_rows,
-                    column_node.field.clone(),
-                )?)
-            }
-        }
-
-        if nested_fields.is_empty() {
-            let mut deserializer =
-                RowGroupDeserializer::new(columns_array_iter, part.num_rows, None);
-            return self.full_deserialize(&mut deserializer);
-        }
-
-        let bitmap = filter.unwrap();
-        let normal_block = try_next_block(
-            &DataSchema::new(normal_fields.clone()),
-            &mut RowGroupDeserializer::new(columns_array_iter, part.num_rows, None),
-        )?;
-        let nested_block = try_next_block(
-            &DataSchema::new(nested_fields.clone()),
-            &mut RowGroupDeserializer::new(nested_columns_array_iter, part.num_rows, None),
-        )?;
-        // need to filter nested block
-        let nested_block = DataBlock::filter_with_bitmap(nested_block, &bitmap)?;
-
-        // Construct the final output
-        let mut final_columns = Vec::with_capacity(self.output_schema.fields().len());
-        final_columns.extend_from_slice(normal_block.columns());
-        final_columns.extend_from_slice(nested_block.columns());
-        let final_block = DataBlock::new(final_columns, bitmap.len() - bitmap.unset_bits());
-
-        normal_fields.extend_from_slice(&nested_fields);
-        let src_schema = DataSchema::new(normal_fields);
-        final_block.resort(&src_schema, &self.output_schema)
-    }
-
+impl Parquet2Reader {
     /// The number of columns can be greater than 1 because the it may be a nested type.
     /// Combine multiple columns into one arrow array.
-    fn to_array_iter(
+    pub(crate) fn to_array_iter(
         metas: Vec<(&ColumnMeta, &ColumnDescriptor)>,
         chunks: Vec<Vec<u8>>,
         rows: usize,
@@ -180,7 +80,7 @@ impl ParquetReader {
     }
 
     /// Almost the same as `to_array_iter`, but with a filter.
-    fn to_array_iter_with_filter(
+    pub(crate) fn to_array_iter_with_filter(
         metas: Vec<(&ColumnMeta, &ColumnDescriptor)>,
         chunks: Vec<Vec<u8>>,
         rows: usize,
@@ -248,12 +148,15 @@ impl ParquetReader {
         )?)
     }
 
-    fn full_deserialize(&self, deserializer: &mut RowGroupDeserializer) -> Result<DataBlock> {
+    pub(crate) fn full_deserialize(
+        &self,
+        deserializer: &mut RowGroupDeserializer,
+    ) -> Result<DataBlock> {
         try_next_block(&self.output_schema, deserializer)
     }
 
     // Build a map to record the count number of each leaf_id
-    fn build_projection_count_map(columns: &[ColumnNode]) -> HashMap<usize, usize> {
+    pub(crate) fn build_projection_count_map(columns: &[ColumnNode]) -> HashMap<usize, usize> {
         let mut cnt_map = HashMap::with_capacity(columns.len());
         for column in columns {
             for index in &column.leaf_indices {
@@ -269,7 +172,7 @@ impl ParquetReader {
     }
 }
 
-fn try_next_block(
+pub(crate) fn try_next_block(
     schema: &DataSchema,
     deserializer: &mut RowGroupDeserializer,
 ) -> Result<DataBlock> {
